Move GameView to be a container
