+++ game_frontend/src/containers/GameView/index.test.js
